pyproject.toml
[tool.poetry]
name = "simple-llm-application-with-lcel"
version = "0.1.0"
description = ""
authors = ["hungpm372 <hungpm372@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.2.6"
python-dotenv = "^1.0.1"
langchain-openai = "^0.1.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
